+++ apps/docs-app/src/app/content/components/component-demos/dynamic-menu/demos/dynamic-menu-demo-basic/dynamic-menu-demo-basic.component.ts
@@ -70,7 +70,7 @@ export class DynamicMenuDemoBasicComponent {
         {
           // URL link
           text: 'Angular Homepage',
-          link: 'https://angular.io/',
+          link: 'https://angular.dev/',
           newTab: true,
         },
       ],