+++ libs/components-react/src/ActionRibbon/CovalentActionRibbon.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentActionRibbon } from './CovalentActionRibbon';
+
+describe('CovalentActionRibbon', () => {
+  it('should work', () => {
+    expect(CovalentActionRibbon).toBeTruthy();
+  });
+});