+++ libs/components-react/src/ExpansionPanel/CovalentExpansionPanelItem.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentExpansionPanelItem } from './CovalentExpansionPanelItem';
+
+describe('CovalentExpansionPanelItem', () => {
+  it('should work', () => {
+    expect(CovalentExpansionPanelItem).toBeTruthy();
+  });
+});