+++ libs/components-react/src/Formfield/CovalentFormfield.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentFormfield } from './CovalentFormfield';
+
+describe('CovalentFormfield', () => {
+  it('should work', () => {
+    expect(CovalentFormfield).toBeTruthy();
+  });
+});