+++ libs/components-react/src/IconButtonToggle/CovalentIconButtonToggle.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentIconButtonToggle } from './CovalentIconButtonToggle';
+
+describe('CovalentIconButtonToggle', () => {
+  it('should work', () => {
+    expect(CovalentIconButtonToggle).toBeTruthy();
+  });
+});