+++ libs/components-react/src/Slider/CovalentSlider.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { createComponent } from '@lit/react';
+import { CovalentSlider as CovalentSliderWeb } from '@covalent/components';
+export const CovalentSlider = createComponent({
+  tagName: 'cv-slider',
+  elementClass: CovalentSliderWeb,
+  react: React,
+  events: {
+    oninput: 'input',
+    onchange: 'change',
+  },
+});