+++ libs/components-react/src/StatusHeader/CovalentStatusHeaderItem.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentStatusHeaderItem } from './CovalentStatusHeaderItem';
+
+describe('CovalentStatusHeaderItem', () => {
+  it('should work', () => {
+    expect(CovalentStatusHeaderItem).toBeTruthy();
+  });
+});