+++ libs/components-react/src/TextLockup/CovalentTextLockup.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentTextLockup } from './CovalentTextLockup';
+
+describe('CovalentTextLockup', () => {
+  it('should work', () => {
+    expect(CovalentTextLockup).toBeTruthy();
+  });
+});