+++ libs/components-react/src/TreeList/CovalentTreeListItem.spec.ts
@@ -1,0 +1,7 @@
+import { CovalentTreeListItem } from './CovalentTreeListItem';
+
+describe('CovalentTreeListItem', () => {
+  it('should work', () => {
+    expect(CovalentTreeListItem).toBeTruthy();
+  });
+});